+++ LICENSE.md
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright (c) 2022 OSLabs Beta
+Copyright (c) 2023 OSLabs Beta
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
