+++ extension/chrome/manifest.json
@@ -1,5 +1,5 @@
 {
-  "version": "1.1",
+  "version": "1.2",
   "name": "Zukeeper: Devtools for Zustand",
   "description": "Devtools for Zustand, in Zustand",
   "manifest_version": 3,
