--- src/client/MediaQueries.scss
@@ -1,3 +1,0 @@
-@media (prefers-color-scheme: dark) {
-  @import "../ColorsDarkMode.scss";
-}
