+++ src/client/components/Action/Action.scss
@@ -1,4 +1,4 @@
-@import "./_Themes.scss";
+@import "../../styles/Themes";
 
 .action-button-container {
   display: flex;
