+++ src/client/components/Action/types.ts
@@ -1,0 +1,5 @@
+export type actionProps = {
+  action: string;
+  idx: number;
+  length: number;
+};
