+++ src/client/components/ActionsDispatched/ActionsDispatched.scss
@@ -1,4 +1,4 @@
-@import "./_Themes.scss";
+@import "../../styles/Themes";
 
 .actions-container {
   flex-basis: 30%;
