+++ src/client/components/StateDisplay/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import useExtensionStore from "../store/useExtensionStore";
+import useExtensionStore from "../../store/useExtensionStore";
 import { useStore } from "zustand";
-import "../styles/StateDisplay.scss";
+import "./StateDisplay.scss";
 
 export const StateDisplay = (): JSX.Element => {
   const { previousStates, actionIndex, isDarkMode } =
