+++ src/client/components/TreeDisplay/TreeDisplay.scss
@@ -1,4 +1,4 @@
-@import "./_Themes.scss";
+@import "../../styles/Themes";
 
 .node__root circle {
   fill: var(--tertiary-dark);
