+++ src/client/components/VisualizationContainer/VisualizationContainer.scss
@@ -1,4 +1,4 @@
-@import "./_Themes.scss";
+@import "../../styles/Themes";
 
 #visualization-container {
   flex-basis: 70%;
