+++ src/client/components/VisualizationSelector/VisualizationSelector.scss
@@ -1,4 +1,4 @@
-@import "./_Themes.scss";
+@import "../../styles/Themes";
 
 .visualization-wrapper {
   display: flex;
