+++ src/client/store/types.ts
@@ -1,4 +1,4 @@
-export interface Store {
+export type Store = {
   displayState: boolean;
   displayDiff: boolean;
   showState: () => void;
@@ -41,23 +41,3 @@ export interface Store {
   reset: boolean;
   setReset: (bool: boolean) => void;
 }
-
-export type diffProps = {
-  obj: diffItemTypes;
-  action: boolean;
-};
-
-export interface diffItemTypes {
-  kind: string | undefined;
-  path: any;
-  lhs?: string | undefined;
-  rhs?: string | undefined;
-  index?: number | undefined;
-  item?: any;
-}
-
-export type actionProps = {
-  action: string;
-  idx: number;
-  length: number;
-};
